room api treatment
