+++ remoteMedi/authenticate/urls.py
@@ -4,8 +4,6 @@ from .views import kakaoauth
 from .views import kakaologout
 from .views import redirect_test
 from .views import kakaosendmsg
-from .views import createroom
-from .views import joinroom
 from .views import login
 
 urlpatterns = [
@@ -15,6 +13,4 @@ urlpatterns = [
     path('redirect/test', redirect_test),
     path('kakaosendmsg/', kakaosendmsg),
     path('login/', login),
-    path('api/createroom', createroom),
-    path('api/joinroom', joinroom)
 ]
