+++ remoteMedi/authenticate/views.py
@@ -7,7 +7,6 @@ from django.shortcuts import redirect
 import requests
 from django.conf import settings
 import json
-import hashlib 
 
 # Create your views here.
 
@@ -101,40 +100,3 @@ def kakaosendmsg(request):
     result = requests.post(url=send_url, headers = headers, data = data)
 
     return Response(result)
-
-
-@api_view(['POST'])
-def createroom(request):
-    res = json.loads(request.body)
-    print(res['identity'])
-    
-    enc = hashlib.md5()
-    enc.update(res['identity'].encode('utf-8'))
-    encText = enc.hexdigest()
-    print(encText)
-
-    
-    """
-    #대충 DB관련 중복검사 및 insert 
-    #
-    #
-    #
-    """ 
-    
-    return Response(status=status.HTTP_200_OK)
-    
-@api_view(['POST'])
-def joinroom(request):
-    res = json.loads(request.body)
-    print(res['identity'])    
-    
-    
-    """
-    #대충 DB관련 데이터 검사 select
-    #
-    #
-    #
-    """        
-    roomnumber = 123
-    return Response(roomnumber,status=status.HTTP_200_OK)
-
