+++ remoteMedi/remoteMedi/urls.py
@@ -22,7 +22,8 @@ schema_view = get_schema_view(
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('authenticate/', include('authenticate.urls')),
-    path('healthcheck/', include('healthcheck.urls'))
+    path('healthcheck/', include('healthcheck.urls')),
+    path('treatment/', include('treatment.urls'))
 
 ]
